Ensure the ssh_key exists
Prior to this commit, if the ssh_key was not a file on disk the destroy
would fail. This would happen durring the autope::destroy plan as it did
no pass the ssh_key parameters and uses the default for the variable.
This commit adds `/dev/null` as the default public key in
the case that the ssh_key does not exist on disk.